requirements.txt
pandas==2.1.3
Faker==37.0.1
ipython==8.18.0
google-cloud-bigquery==3.30.0
streamlit==1.43.2
plotly==6.0.1
db-dtypes==1.4.2


